Add files via upload
# Project description
This project contains code in Wolfram Mathematica to calculate the utility function of three modes of transportation from linear demands with respect to generalized costs. Additionally, the process of obtaining the direct and inverse demand parameters.


## Results

The code produces the direct and inverse demand functions with their respective relationships between parameters for subsequent estimation.

## Contributions

Contributions are welcome. If you encounter any problems or want to add new features, please create an issue or submit a pull request.
coefficient