Add files via upload
This is an Excel file containing the parameters of the model used in an academic article currently in the process of publication. We develop a microeconomic model of competition between ridesourcing, public transport, and cruising taxis to analyse how alternative ridesourcing regulation policies affect not only the ridesourcing service but also other modes of transport. Full details of the model and its application can be found in the article titled "Microeconomic analysis of ridesourcing market regulation policies", which is currently in the process of publication.